requirements.txt
streamlit>=1.32.0
groq>=0.5.0
neo4j>=5.17.0
pinecone-client>=3.1.0
sentence-transformers>=2.6.1
langchain>=0.1.13
langchain-community>=0.0.28
langchain-groq>=0.0.1
plotly>=5.20.0
pyvis>=0.3.2
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
